+++ apps/web/src/entities/feed/index.ts
@@ -1,0 +1,6 @@
+export type { FeedPage, FeedPost, FeedTag } from './model/types';
+export {
+  FEED_QUERY_KEY,
+  useInfiniteFeedQuery,
+} from './model/useInfiniteFeedQuery';
+export { default as FeedCard } from './ui/FeedCard';
