+++ apps/web/src/features/scroll-to-top/index.ts
@@ -1,0 +1,1 @@
+export { default as ScrollToTopButton } from './ui/ScrollToTopButton';
